Refactor 'HostingSection' to 'PageSection' and convert it into a global component.
