+++ client/a8c-for-agencies/components/page-section/style.scss
@@ -1,7 +1,7 @@
 @import "@wordpress/base-styles/breakpoints";
 @import "@wordpress/base-styles/mixins";
 
-.hosting-section-wrapper {
+.page-section-wrapper {
 	margin-inline: 0;
 	background-repeat: no-repeat;
 	background-position: top right;
@@ -22,18 +22,18 @@
 	}
 }
 
-.hosting-section__sub-header {
+.page-section__sub-header {
 	display: flex;
 	flex-direction: row;
 	gap: 8px;
 }
 
-.hosting-section__sub-header-title {
+.page-section__sub-header-title {
 	@include a4a-font-body-md;
 }
 
 
-.hosting-section__header {
+.page-section__header {
 	display: flex;
 	flex-direction: column;
 	gap: 8px;
@@ -44,7 +44,7 @@
 	}
 }
 
-.hosting-section__header-title {
+.page-section__header-title {
 	@include a4a-font-heading-xl;
 
 	@include break-medium {
@@ -52,7 +52,7 @@
 	}
 }
 
-.hosting-section__header-description {
+.page-section__header-description {
 	@include a4a-font-body-md;
 	margin: 0;
