+++ client/a8c-for-agencies/sections/marketplace/hosting-overview/hosting-v2/enterprise-agency-hosting/index.tsx
@@ -2,6 +2,7 @@ import { VIPLogo, SalesforceLogo, SlackLogo, CNNLogo } from '@automattic/compone
 import { Button } from '@wordpress/components';
 import { Icon, external } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
+import { BackgroundType4 } from 'calypso/a8c-for-agencies/components/page-section/backgrounds';
 import ProfileAvatar1 from 'calypso/assets/images/a8c-for-agencies/hosting/enterprise-testimonial-1.png';
 import ProfileAvatar2 from 'calypso/assets/images/a8c-for-agencies/hosting/enterprise-testimonial-2.png';
 import CapeGeminiLogo from 'calypso/assets/images/logos/capgemini.svg';
@@ -12,7 +13,6 @@ import SpotifyLogo from 'calypso/assets/images/logos/spotify.svg';
 import { useDispatch } from 'calypso/state';
 import { recordTracksEvent } from 'calypso/state/analytics/actions';
 import HostingAdditionalFeaturesSection from '../../../common/hosting-additional-features-section';
-import { BackgroundType4 } from '../../../common/hosting-section/backgrounds';
 import HostingTestimonialsSection from '../../../common/hosting-testimonials-section';
 import SimpleList from '../../../common/simple-list';
